+++ packages-next/fields-document/src/DocumentEditor/pasting/data-transfer.ts
