+++ packages-next/fields-document/src/DocumentEditor/pasting/pasting-from-other-slate-editor.test.tsx
@@ -4,7 +4,7 @@ import { createEditor, Editor } from 'slate';
 import { Editable, ReactEditor, Slate, withReact } from 'slate-react';
 import { render } from '@testing-library/react';
 import { makeEditor } from '../tests/utils';
-import { MyDataTransfer } from './test-utils';
+import { MyDataTransfer } from './data-transfer';
 
 function OtherEditor({ editor }: { editor: ReactEditor }) {
   // note that the entire point here is to have a different document structure to Keystone's
